Moved fetchCommand method to messages
